refactor mentee profile
